refactor: move editor components to components/editor
Create folders for components/puck and components/puck/atoms

Testing on dev account
