+++ README.md
@@ -14,9 +14,9 @@ This library provides components necessary to set up a Pages repository that can
 
 | Component                                             |
 |-------------------------------------------------------|
-| [Editor](./src/components/README.md#editor)           |
-| [EntityField](./src/components/README.md#entityfield) | 
-| [YextEntityFieldSelector](./src/components/README.md#YextEntityFieldSelector) | 
+| [Editor](src/editor/README.md#editor)           |
+| [EntityField](src/editor/README.md#entityfield) | 
+| [YextEntityFieldSelector](src/editor/README.md#YextEntityFieldSelector) | 
 
 ## Utils
