+++ src/components/editor/README.md
