+++ src/components/editor/index.ts
