+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from "./components/index.ts";
+export * from "./components/editor/index.ts";
 export * from "./hooks/index.ts";
 export * from "./utils/index.js";
