+++ src/internal/puck/components/LoadingScreen.tsx
@@ -1,7 +1,7 @@
 import { Progress } from "../ui/Progress.tsx";
 import React from "react";
 import "../ui/puck.css";
-import "../../../components/index.css";
+import "../../../components/editor/index.css";
 
 export type LoadingScreenProps = {
   progress: number;
